Simplify resolution of the application's main class name
Closes gh-30467
