+++ spring-boot-project/spring-boot-tools/spring-boot-gradle-plugin/src/main/java/org/springframework/boot/gradle/plugin/JavaPluginAction.java
@@ -35,6 +35,8 @@ import org.gradle.api.file.FileCollection;
 import org.gradle.api.model.ObjectFactory;
 import org.gradle.api.plugins.ApplicationPlugin;
 import org.gradle.api.plugins.BasePlugin;
+import org.gradle.api.plugins.ExtensionContainer;
+import org.gradle.api.plugins.JavaApplication;
 import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.JavaPluginExtension;
 import org.gradle.api.provider.Provider;
@@ -46,6 +48,7 @@ import org.gradle.api.tasks.compile.JavaCompile;
 import org.gradle.jvm.toolchain.JavaToolchainService;
 import org.gradle.jvm.toolchain.JavaToolchainSpec;
 
+import org.springframework.boot.gradle.dsl.SpringBootExtension;
 import org.springframework.boot.gradle.tasks.bundling.BootBuildImage;
 import org.springframework.boot.gradle.tasks.bundling.BootJar;
 import org.springframework.boot.gradle.tasks.run.BootRun;
@@ -77,10 +80,11 @@ final class JavaPluginAction implements PluginApplicationAction {
 		classifyJarTask(project);
 		configureBuildTask(project);
 		configureDevelopmentOnlyConfiguration(project);
-		TaskProvider<BootJar> bootJar = configureBootJarTask(project);
+		TaskProvider<ResolveMainClassName> resolveMainClassName = configureResolveMainClassNameTask(project);
+		TaskProvider<BootJar> bootJar = configureBootJarTask(project, resolveMainClassName);
 		configureBootBuildImageTask(project, bootJar);
 		configureArtifactPublication(bootJar);
-		configureBootRunTask(project);
+		configureBootRunTask(project, resolveMainClassName);
 		project.afterEvaluate(this::configureUtf8Encoding);
 		configureParametersCompilerArg(project);
 		configureAdditionalMetadataLocations(project);
@@ -96,7 +100,39 @@ final class JavaPluginAction implements PluginApplicationAction {
 				.configure((task) -> task.dependsOn(this.singlePublishedArtifact));
 	}
 
-	private TaskProvider<BootJar> configureBootJarTask(Project project) {
+	private TaskProvider<ResolveMainClassName> configureResolveMainClassNameTask(Project project) {
+		return project.getTasks().register(SpringBootPlugin.RESOLVE_MAIN_CLASS_NAME_TASK_NAME,
+				ResolveMainClassName.class, (resolveMainClassName) -> {
+					ExtensionContainer extensions = project.getExtensions();
+					resolveMainClassName.setDescription("Resolves the name of the application's main class.");
+					resolveMainClassName.setGroup(BasePlugin.BUILD_GROUP);
+					Callable<FileCollection> classpath = () -> project.getExtensions()
+							.getByType(SourceSetContainer.class).getByName(SourceSet.MAIN_SOURCE_SET_NAME).getOutput();
+					resolveMainClassName.setClasspath(classpath);
+					resolveMainClassName.getConfiguredMainClassName().convention(project.provider(() -> {
+						String javaApplicationMainClass = getJavaApplicationMainClass(extensions);
+						if (javaApplicationMainClass != null) {
+							return javaApplicationMainClass;
+						}
+						SpringBootExtension springBootExtension = project.getExtensions()
+								.findByType(SpringBootExtension.class);
+						return springBootExtension.getMainClass().getOrNull();
+					}));
+					resolveMainClassName.getOutputFile()
+							.set(project.getLayout().getBuildDirectory().file("resolvedMainClassName"));
+				});
+	}
+
+	private static String getJavaApplicationMainClass(ExtensionContainer extensions) {
+		JavaApplication javaApplication = extensions.findByType(JavaApplication.class);
+		if (javaApplication == null) {
+			return null;
+		}
+		return javaApplication.getMainClass().getOrNull();
+	}
+
+	private TaskProvider<BootJar> configureBootJarTask(Project project,
+			TaskProvider<ResolveMainClassName> resolveMainClassName) {
 		SourceSet mainSourceSet = javaPluginExtension(project).getSourceSets()
 				.getByName(SourceSet.MAIN_SOURCE_SET_NAME);
 		Configuration developmentOnly = project.getConfigurations()
@@ -105,8 +141,6 @@ final class JavaPluginAction implements PluginApplicationAction {
 				.getByName(SpringBootPlugin.PRODUCTION_RUNTIME_CLASSPATH_CONFIGURATION_NAME);
 		Callable<FileCollection> classpath = () -> mainSourceSet.getRuntimeClasspath()
 				.minus((developmentOnly.minus(productionRuntimeClasspath))).filter(new JarTypeFileSpec());
-		TaskProvider<ResolveMainClassName> resolveMainClassName = ResolveMainClassName
-				.registerForTask(SpringBootPlugin.BOOT_JAR_TASK_NAME, project, classpath);
 		return project.getTasks().register(SpringBootPlugin.BOOT_JAR_TASK_NAME, BootJar.class, (bootJar) -> {
 			bootJar.setDescription(
 					"Assembles an executable jar archive containing the main classes and their dependencies.");
@@ -134,11 +168,9 @@ final class JavaPluginAction implements PluginApplicationAction {
 		this.singlePublishedArtifact.addJarCandidate(bootJar);
 	}
 
-	private void configureBootRunTask(Project project) {
+	private void configureBootRunTask(Project project, TaskProvider<ResolveMainClassName> resolveMainClassName) {
 		Callable<FileCollection> classpath = () -> javaPluginExtension(project).getSourceSets()
 				.findByName(SourceSet.MAIN_SOURCE_SET_NAME).getRuntimeClasspath().filter(new JarTypeFileSpec());
-		TaskProvider<ResolveMainClassName> resolveProvider = ResolveMainClassName.registerForTask("bootRun", project,
-				classpath);
 		project.getTasks().register("bootRun", BootRun.class, (run) -> {
 			run.setDescription("Runs this project as a Spring Boot application.");
 			run.setGroup(ApplicationPlugin.APPLICATION_GROUP);
@@ -149,7 +181,7 @@ final class JavaPluginAction implements PluginApplicationAction {
 				}
 				return Collections.emptyList();
 			});
-			run.getMainClass().convention(resolveProvider.flatMap(ResolveMainClassName::readMainClassName));
+			run.getMainClass().convention(resolveMainClassName.flatMap(ResolveMainClassName::readMainClassName));
 			configureToolchainConvention(project, run);
 		});
 	}
